Add authentication, authorization and tests

+++ part4/blog/controllers/blogs.js
@@ -1,20 +1,26 @@
 import Blog from '../models/blog.js';
+import User from '../models/user.js';
 
 export const getAllBlogs = async (_request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
   return response.json(blogs);
 };
 
 export const createBlog = async (request, response) => {
-  const { title, url } = request.body;
+  const { author, title, url } = request.body;
 
   if (!title || !url) {
     return response.status(400).json({ error: 'Please provide title and url' });
   }
 
-  const blog = new Blog(request.body);
+  const user = await User.findById(request.user.id);
+
+  const blog = new Blog({ author, title, url, user: user.id });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 };
 
@@ -40,6 +46,13 @@ export const deleteBlog = async (request, response) => {
 
   if (!existingBlog) {
     return response.status(404).json({ error: `No blog with id ${id}` });
+  }
+
+  console.log(existingBlog);
+  console.log(request.user);
+
+  if (existingBlog.user._id.toString() !== request.user.id.toString()) {
+    return response.status(404).json({ error: `Permission denied` });
   }
 
   await Blog.findByIdAndDelete(id);
